requirements.txt
Django==1.7.3
psycopg2==2.5.4
django-haystack==2.3.1
pyelasticsearch==1.0
djangorestframework==3.0.3
